+++ src/main/java/com/example/mentoring/order/OrderService.java
@@ -1,6 +1,6 @@
 package com.example.mentoring.order;
 
-import com.example.mentoring.common.Timer;
+import com.example.mentoring.common.aop.Timer;
 import com.example.mentoring.exception.domain.MenuNotFoundException;
 import com.example.mentoring.food.domain.Food;
 import com.example.mentoring.food.domain.FoodManager;
